+++ inst/examples/dag_targets/first_target.R
@@ -1,0 +1,4 @@
+dir.create("output", showWarnings=FALSE)
+
+message("Writing output/first_target.dat")
+write(2, "output/first_target.dat")