+++ inst/examples/dag_targets/global_object.R
@@ -1,0 +1,4 @@
+dir.create("output", showWarnings=FALSE)
+
+message("Writing output/global_object")
+write(3, "output/global_object.dat")