+++ inst/examples/dag_wikipedia/a.R
@@ -1,0 +1,4 @@
+dir.create("out", showWarnings=FALSE)
+
+message("Writing out/a.dat")
+write(utf8ToInt("a"), "out/a.dat")