+++ inst/examples/dag_wikipedia/b.R
@@ -1,0 +1,4 @@
+a <- scan("out/a.dat", quiet=TRUE)
+
+message("Writing out/b.dat")
+write(a + utf8ToInt("b"), "out/b.dat")