+++ NEWS.md
@@ -1,3 +1,16 @@
+# makeit 1.0.1 (2023-10-27)
+
+* Renamed argument 'silent' to 'quiet' in make().
+
+* Improved error messages when make() is called using invalid 'recipe',
+  'prereq', or 'target'.
+
+* Added a tutorial vignette with annotated examples and discussion, see
+  vignette("makeit").
+
+
+
+
 # makeit 1.0.0 (2023-03-28)
 
 * Initial release, providing the the function make().